+++ lib/repo/fetch.js
@@ -1,0 +1,97 @@
+var Promise = require( "es6-promise" ).Promise,
+	fs = require( "fs" ),
+	exec = require( "child_process" ).exec,
+	mkdirp = require( "mkdirp" );
+
+module.exports = function( Repo ) {
+
+Repo.prototype.fetchBranch = function( branch ) {
+	return this.fetch( "heads/" + branch );
+};
+
+Repo.prototype.fetchPr = function( pr ) {
+	return this.fetch( "pull/" + pr + "/head" );
+};
+
+Repo.prototype.fetch = function( ref ) {
+	var repo = this;
+	return repo.exists()
+		.then(function( exists ) {
+			if ( !exists ) {
+				return repo._clone();
+			}
+		})
+		.then(function() {
+			return repo._fetch( ref );
+		});
+};
+
+Repo.prototype.exists = function() {
+	var repo = this;
+	return new Promise(function( resolve, reject ) {
+		fs.stat( repo.path, function( error ) {
+
+			// Repo already exists
+			if ( !error ) {
+				repo.debug( "local repo already exists" );
+				return resolve( true );
+			}
+
+			// Error other than repo not existing
+			if ( error.code !== "ENOENT" ) {
+				repo.debug( "error checking if a local repo exists", error );
+				return reject( error );
+			}
+
+			repo.debug( "local repo does not yet exist" );
+			resolve( false );
+		});
+	});
+};
+
+Repo.prototype._clone = function() {
+	var repo = this;
+
+	repo.debug( "cloning repo" );
+	return new Promise(function( resolve, reject ) {
+
+		// Ensure the path exists before trying to clone
+		mkdirp( repo.path, "0755", function( error ) {
+			if ( error ) {
+				repo.debug( "error creating path", error );
+				return reject( error );
+			}
+
+			exec( "git clone " + repo.remoteUrl + " " + repo.path, function( error ) {
+				if ( error ) {
+					repo.debug( "error cloning " + repo.remoteUrl, error );
+					return reject( error );
+				}
+
+				repo.debug( "successfully cloned repo" );
+				resolve();
+			});
+		});
+	});
+};
+
+Repo.prototype._fetch = function( ref ) {
+	var repo = this;
+
+	repo.debug( "fetching repo" );
+	return new Promise(function( resolve, reject ) {
+		exec( "git fetch -fu origin refs/" + ref + ":" + ref, {
+			cwd: repo.path
+		}, function( error ) {
+			if ( error ) {
+				repo.debug( "error fetching repo", error );
+				return reject( error );
+			}
+
+			repo.debug( "successfully fetched repo" );
+			resolve();
+		});
+	});
+};
+
+};
